Refine unit test for trace.db

+++ edu.rice.cs.hpcdata/src/edu/rice/cs/hpcdata/db/version4/DataTrace.java
@@ -168,7 +168,7 @@ public class DataTrace extends DataCommon
 	 */
 	public int getNumberOfRanks()
 	{
-		return (int) traceCtxs.length;
+		return traceCtxs.length;
 	}
